requirements.txt
pillow
bitstring
numpy
scip
